substr

+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: lpittet <lpittet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/03 11:04:58 by lpittet           #+#    #+#             */
-/*   Updated: 2024/10/07 16:45:51 by lpittet          ###   ########.fr       */
+/*   Updated: 2024/10/07 19:44:57 by lpittet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char			*sub;
-	unsigned long	i;
+	size_t			i;
 
 	i = 0;
 	sub = malloc((len + 1) * sizeof(char));
